move `ToolbarHelper` to `Joomla\CMS\Toolbar`
